[CI] [57259]  <multimedia> <feat> opus:support prompt tone resample

+++ components/multimedia/prompt_tone_play/src/mp3_codec.c
@@ -793,7 +793,8 @@ static int mp3_codec_write(audio_codec_t *codec, char *buffer, uint32_t len)
     mp3_codec_priv_t *priv = (mp3_codec_priv_t *)codec->codec_ctx;
     MP3_CODEC_CHECK_NULL(priv);
 
-    uint32_t need_w_len = len;
+    //uint32_t need_w_len = len;
+    int need_w_len = (int)len;
 
     while (need_w_len)
     {
